% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertId}
\alias{convertId}
\title{convertId}
\usage{
convertId(x, dataset = "hsapiens_gene_ensembl",
  filters = "uniprot_swissprot_accession", attributes = c(filters,
  "entrezgene"), genesKept = c("foldchange", "first", "random", "var", "abs"),
  keepNoId = T, keepMultipleId = F, verbose = F)
}
\arguments{
\item{x}{the expression data matrix.}

\item{dataset}{Dataset you want to use.  To see the different datasets available within a biomaRt you can e.g. do:  mart = useMart('ensembl'), followed by listDatasets(mart).}

\item{filters}{Filters (one or more) that should be used in the query.  A possible list of filters can be retrieved using the function listFilters.}

\item{attributes}{Attributes you want to retrieve.  A possible list of attributes can be retrieved using the function listAttributes.}

\item{genesKept}{The method to select target gene in more than one targets. "var"/"foldchange"/"abs" means selecting the gene with largest variation/fold change/absolute value. "first" means selecting the first target and "random" means randomly selection.}

\item{keepNoId}{Logical. Indicate keep the source IDs without target IDs or not.}

\item{keepMultipleId}{Logical. Indicate keep the multiple target IDs related to one source ID or not.}

\item{verbose}{Logical. Indicate report extra information on progress or not.}
}
\description{
A function to convert ID based on the biomaRt package.
}
\details{
A function to convert ID based on the biomaRt package..
}
\examples{
temp<-cbind(rnorm(10),rnorm(10))
row.names(temp)<-c("Q04837","P0C0L4","P0C0L5","O75379","Q13068","A2MYD1","P60709","P30462","P30475","P30479")
colnames(temp)<-c("Exp1","Exp2")
convertId(temp,filters="uniprot_swissprot",keepMultipleId=TRUE)
\dontrun{
temp<-cbind(rnorm(5000),rnorm(5000),rnorm(5000),rnorm(5000),rnorm(5000),rnorm(5000))
row.names(temp)<-1000:5999
colnames(temp)<-c("Control1","Control2","Control3","Treatment1","Treatment2","Treatment3")
convertId(temp,filters="entrezgene",attributes =c("entrezgene","uniprot_swissprot"),keepNoId=FALSE)
}
}

