% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{download_latest_pathway}
\alias{download_latest_pathway}
\title{download_latest_pathway}
\usage{
download_latest_pathway(species)
}
\arguments{
\item{species}{the species id in KEGG database, 'hsa' means human, 'mmu' means mouse, 'rno' means rat, etc}
}
\value{
a list with two parts \item{name keggpathway2gene}{description a list with the genes for each pathway}\item{name pathway2name}{description a list with the names for each pathway}
}
\description{
The function will downlaod the latest pathway gene link from KEGG website.
}
\details{
The function will downlaod the latest pathway gene link from KEGG website.
}
\examples{
\dontrun{download_latest_pathway(species="hsa")}
}

