% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGprofile.R
\docType{package}
\name{KEGGprofile-package}
\alias{KEGGprofile}
\alias{KEGGprofile-package}
\title{KEGGprofile: An annotation and visualization package for multi-types and
    multi-groups expression data in KEGG pathway}
\description{
KEGGprofile is an annotation and visualization tool which
    integrated the expression profiles and the function annotation in KEGG
    pathway maps. The multi-types and multi-groups expression data can be
    visualized in one pathway map. KEGGprofile facilitated more detailed
    analysis about the specific function changes inner pathway or temporal
    correlations in different genes and samples.
}
