% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{pho_sites_count}
\alias{pho_sites_count}
\title{number of phosphorylation sites quantified for each gene}
\source{
Olsen, J.V., et al. (2010) Quantitative phosphoproteomics reveals
widespread full phosphorylation site occupancy during mitosis, Sci Signal, 3, ra3.
}
\usage{
pho_sites_count
}
\description{
This data set is a data.frame with number of phosphorylation sites quantified for each gene in the analysis.
}

