% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_function.R
\name{plot_pathway_cor}
\alias{plot_pathway_cor}
\title{plot_pathway_cor}
\usage{
plot_pathway_cor(gene_expr, kegg_enriched_pathway, groups = NULL,
  side = c("both", "pos", "neg"), alternative = NULL)
}
\arguments{
\item{gene_expr}{the matrix for gene expression, row.names should be NCBI gene ID, such as 67040, 93683}

\item{kegg_enriched_pathway}{The returned value from find_enriched_pathway function, the enriched pathways.}

\item{groups}{a character used to indicate expression values from different types of samples}

\item{side}{a character string specifying the correlation directions interested, must be one of "both" (default), "pos" or "neg".}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
p values for Wilcoxon tests in each pathway
}
\description{
The function will plot the correlation distributions for each enriched pathway (result from find_enriched_pathway function), and then Wilcoxon tests would be used to estimate the significance of correlations distribution between genes in each pathway and all genes.
}
\examples{
data(pro_pho_expr)
data(pho_sites_count)
genes<-row.names(pho_sites_count)[which(pho_sites_count>=10)]
pho_KEGGresult<-find_enriched_pathway(genes,species='hsa')
result<-plot_pathway_cor(gene_expr=pro_pho_expr,kegg_enriched_pathway=pho_KEGGresult)
}
