% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_enriched_pathway}
\alias{find_enriched_pathway}
\title{find_enriched_pathway}
\usage{
find_enriched_pathway(gene, species = "hsa", returned_pvalue = 0.01,
  returned_adjpvalue = 0.05, returned_genenumber = 5,
  download_latest = FALSE)
}
\arguments{
\item{gene}{a numeric matrix}

\item{species}{the species id in KEGG database, 'hsa' means human, 'mmu' means mouse, 'rno' means rat, etc}

\item{returned_pvalue}{the minimum p value for enriched pathways}

\item{returned_adjpvalue}{the minimum adjusted p value for enriched pathways}

\item{returned_genenumber}{the minimum number of annotated genes for enriched pathways}

\item{download_latest}{logical. Indicate if the function will download the latest pathway/gene link from KEGG website. As the KEGG.db package was not updated for a long time due to the KEGG policy change, we provided this parameter so that the users could get the latest KEGG database.}
}
\value{
a list with two parts \item{name stastic}{description a matirx containing the pathway IDs of enriched pathways, and their names, p values, number of annotated genes}\item{name detail}{description a list with the genes annotated for each pathway}
}
\description{
The function will map the genes in KEGG pathway database, and then hypergegeometric tests would be used to estimate the significance of enrichment for each pathway
}
\details{
Only the pathways with p value <= returned_pvalue in hypergegeometric tests and number of annotated genes >= returned_genenumber would be taken as enriched and returned.
}
\examples{
data(pho_sites_count)
#the 300 genes with most phospholation sites quantified
genes<-names(rev(sort(pho_sites_count[,1]))[1:300])
pho_KEGGresult<-find_enriched_pathway(genes,species='hsa')
}

