% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_overall.R
\name{plot_pathway_overall}
\alias{plot_pathway_overall}
\title{plot_pathway_overall}
\usage{
plot_pathway_overall(gene_values, species = "hsa",
  pathwayNumInFigure = 5, rankByVar = colnames(gene_values)[1])
}
\arguments{
\item{gene_values}{A data.frame or matrix with gene ID as rownames. Each column represent gene value in one condition.}

\item{species}{the species id in KEGG database, 'hsa' means human, 'mmu' means mouse, 'rno' means rat, etc}

\item{pathwayNumInFigure}{a integer specifying max number of (top) pathways in each direction in the boxplot.}

\item{rankByVar}{a character string specifying variable (sample) name by which to rank the boxplot.}
}
\value{
p values for Wilcoxon tests in each pathway
}
\description{
The function will plot the correlation distributions for each enriched pathway (result from find_enriched_pathway
 function), and then Wilcoxon tests would be used to estimate the significance of values distribution between 
 genes in each pathway and all other genes.
}
\examples{
data(pro_pho_expr)
data(pho_sites_count)
gene_values<-pro_pho_expr[row.names(pho_sites_count)[which(pho_sites_count>=10)],]
plot_pathway_overall(gene_values=gene_values[,1:3])
}
