% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_overall.R
\name{plot_pathway_overall}
\alias{plot_pathway_overall}
\title{plot_pathway_overall}
\usage{
plot_pathway_overall(gene_values, kegg_enriched_pathway, groups = NULL,
  side = c("both", "pos", "neg"), alternative = NULL)
}
\arguments{
\item{gene_values}{A data.frame or matrix with gene ID as rownames. Each column represent gene value in one condition.}

\item{kegg_enriched_pathway}{The returned value from find_enriched_pathway function, the enriched pathways.}

\item{groups}{a character used to indicate expression values from different types of samples}

\item{side}{a character string specifying interested test directions, must be one of "both" (default), "pos" or "neg".}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
p values for Wilcoxon tests in each pathway
}
\description{
The function will plot the correlation distributions for each enriched pathway (result from find_enriched_pathway
 function), and then Wilcoxon tests would be used to estimate the significance of values distribution between 
 genes in each pathway and all other genes.
}
\examples{
data(pro_pho_expr)
data(pho_sites_count)
genes<-row.names(pho_sites_count)[which(pho_sites_count>=10)]
pho_KEGGresult<-find_enriched_pathway(genes,species='hsa',returned_genenumber=3)
plot_pathway_overall(gene_values=pro_pho_expr[,1:3],kegg_enriched_pathway=pho_KEGGresult)
}
