% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_profile}
\alias{plot_profile}
\title{plot_profile}
\usage{
plot_profile(gene_expr, pathway_name, result_name = paste(pathway_name,
  "_profile_", type, ".png", sep = ""), KEGG_database, groups,
  bg_col = "white", text_col = "black", line_col, border_col = "grey",
  text_cex = 0.25, magnify = 1, type = c("lines", "bg"),
  pathway_min = 5, genes_kept = c("foldchange", "first", "random", "var",
  "abs"), species = "hsa", database_dir = getwd(), max_dist, lwd = 1.2,
  speciesRefMap = TRUE)
}
\arguments{
\item{gene_expr}{the matrix for gene expression, row.names should be NCBI gene ID, such as 67040, 93683}

\item{pathway_name}{the species id and KEGG pathway id, such as 'hsa00010'}

\item{result_name}{the name of figure file generated by KEGGprofile. The default name is pathway_name+'_profile_'+type+'.png', such as 'hsa04110_profile_lines.png'}

\item{KEGG_database}{the matrix returned by function parse_XMLfile, which contains genes in this pathway, and their names, locations etc}

\item{groups}{a character used to indicate expression values from different types of samples}

\item{bg_col}{background color for gene rectangles in the pathway map}

\item{text_col}{the colors for text in the pathway map. A color matrix generated by function \code{\link{col_by_value}} can be used here}

\item{line_col}{line color for expression in different samples in the pathway map, valid when type='lines'}

\item{border_col}{border color for gene rectangles in the pathway map. A color matrix generated by function \code{\link{col_by_value}} can be used here}

\item{text_cex}{cex for text in the pathway map. A color matrix generated by function \code{\link{col_by_value}} can be used here}

\item{magnify}{the coefficient used to magnify the gene rectangles}

\item{type}{the type of pathway map visulization, could be 'bg' or 'lines'. Default is 'bg'. See also 'Details'}

\item{pathway_min}{The pathways with number of annotated genes less than pathway_min would be ignored}

\item{genes_kept}{methods used for choosing genes when several genes corresponding to one location in pathway map. Default is 'foldchange', which kept the gene with largest fold changes. 'first' kept the first gene. 'random' chosed gene random. 'var' kept the gene with largest variation. 'abs' kept the gene with largest absolute value}

\item{species}{the species id in KEGG database, 'hsa' means human, 'mmu' means mouse, 'rno' means rat, etc}

\item{database_dir}{the directory where the XML files and png files are located}

\item{max_dist}{The expression changes that represented by the distance from the bottom to the top of gene rectangle, valid when type='lines'. This param is used to ensure the dynamic changes of lines in different gene polygon represent equal variation. It would be calculated from the maximum changes of genes in this pathway by default. If max_dist=NA, then the lines would be plotted from top to bottom in each gene rectangle}

\item{lwd}{The line width when type='lines'}

\item{speciesRefMap}{Logical, use the species specific figure as reference map. if set as FALSE, the reference pathway figure without species information will be used}
}
\value{
a matrix containing genes maped in this pathway, and their names, expressions
}
\description{
The function plot gene expression profiles on KEGG pathway maps
}
\details{
There are two visualization methods to represent gene expression profiles: 'background' and 'lines'.
The first one is applicable for analysis with only one sample or one type of data, which divides the gene polygon into several sub-polygons to represent different time points.
And each sub-polygon has a specific background color to represent expression changes in that time point. The second method plots lines with different colors in the gene polygon to represent different samples or different types of data.
The dynamic changes of lines mean the profiles of genes in different time points.
}
\examples{
XML2database<-parse_XMLfile(pathway_id="04110",species="hsa",database_dir=system.file("extdata",package="KEGGprofile"))
data(pro_pho_expr)
temp<-plot_profile(pro_pho_expr,pathway_name="hsa04110",KEGG_database=XML2database,line_col=c("brown1","seagreen3"),groups=c(rep("Proteome ",6),rep("Phosphoproteome ",6)),magnify=1.2,database_dir=system.file("extdata",package="KEGGprofile"),max_dist=5)
}

